Refactor mutations and related files
