+++ src/server/mutations/index.ts
@@ -1,0 +1,27 @@
+// Re-export all mutations from domain-specific files
+// This maintains backward compatibility with existing imports
+
+// Import and re-export async functions directly
+import { createGame, updateGameAsFinished, cloneGame } from "./games";
+import { createRoundForGame, updateRoundScores } from "./rounds";
+import {
+  createFriendRequest,
+  acceptFriendRequest,
+  rejectFriendRequest,
+} from "./friends";
+import { createGuestUser, getMyGuestUsers, inviteGuestUser } from "./guests";
+
+// Re-export them
+export {
+  createGame,
+  updateGameAsFinished,
+  cloneGame,
+  createRoundForGame,
+  updateRoundScores,
+  createFriendRequest,
+  acceptFriendRequest,
+  rejectFriendRequest,
+  createGuestUser,
+  getMyGuestUsers,
+  inviteGuestUser,
+};
